OrDo/requirements_learning.txt
# Otto Learning System Dependencies
# =================================

# Core dependencies
PyYAML>=6.0
schedule>=1.2.0

# Optional: For advanced text processing
nltk>=3.8
spacy>=3.5.0

# Optional: For machine learning features
scikit-learn>=1.3.0
numpy>=1.24.0

# Optional: For data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For web interface (future)
flask>=2.3.0
dash>=2.11.0 